requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.17.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
apprise==1.9.6
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgi-lifespan==2.1.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
Authlib==1.6.6
azure-ai-formrecognizer==3.3.3
azure-ai-inference==1.0.0b9
azure-common==1.1.28
azure-core==1.37.0
azure-identity==1.25.1
azure-mgmt-core==1.6.0
azure-storage-blob==12.19.0
azureml-mlflow==1.61.0.post1
babel==2.17.0
backports.tarfile==1.2.0
beartype==0.22.9
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
blosc2==3.12.2
cachetools==6.2.4
catboost==1.2.8
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
coolname==2.2.0
cryptography==45.0.7
curl_cffi==0.13.0
cvxpy==1.7.5
cycler==0.12.1
cyclopts==4.4.4
databricks-sdk==0.76.0
dataclasses-json==0.6.7
dateparser==1.2.2
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fakeredis==2.33.0
fastapi==0.128.0
fastjsonschema==2.21.2
fastmcp==2.14.2
fastuuid==0.14.0
filelock==3.20.2
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
fuzzywuzzy==0.18.0
genai-prices==0.0.49
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.45.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.3.0
griffe==1.15.0
grpcio==1.67.1
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==8.38.0
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.12.0
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaggle==1.7.4.5
keyring==25.7.0
kiwisolver==1.4.9
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
lark==1.3.1
Levenshtein==0.27.3
lightgbm==4.6.0
litellm==1.80.11
logfire-api==4.16.0
loguru==0.7.3
lupa==2.6
lxml==6.0.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.25.0
mdurl==0.1.2
mistune==3.2.0
mlflow==3.5.0
mlflow-skinny==3.5.0
mlflow-tracing==3.5.0
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
msrest==0.7.1
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.14.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.1
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.1
notebook_shim==0.2.4
numexpr==2.14.1
numpy==2.2.6
nvidia-nccl-cu12==2.28.9
oauthlib==3.3.1
openai==2.14.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
osqp==1.0.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandarallel==1.6.5
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathable==0.4.4
pathspec==0.12.1
pathvalidate==3.3.1
peewee==3.18.3
pendulum==3.1.0
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
prefect==3.6.9
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-ai-slim==1.39.0
pydantic-extra-types==2.11.0
pydantic-graph==1.39.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydocket==0.16.3
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
PyMuPDF==1.26.7
pyparsing==3.3.1
pypdf==6.5.0
pyperclip==1.11.0
pyqlib==0.9.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-Levenshtein==0.27.3
python-multipart==0.0.21
python-redis-lock==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
randomname==0.2.1
RapidFuzz==3.14.3
rdagent==0.8.0
readchar==4.2.1
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.15.3
scs==3.2.10
seaborn==0.13.2
SecretStorage==3.5.0
selenium==4.39.0
semver==3.0.4
Send2Trash==2.0.0
setuptools-scm==9.2.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
sse-starlette==3.1.2
st-theme==1.2.3
stack-data==0.6.3
starlette==0.50.0
streamlit==1.52.2
tables==3.10.1
tenacity==9.1.2
termcolor==3.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.25.2
tree-sitter-python==0.25.0
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.1
typer-slim==0.21.0
types-pytz==2025.2.0.20251108
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==1.17.3
wsproto==1.3.2
xgboost==3.1.2
xxhash==3.6.0
yarl==1.22.0
yfinance==1.0
zipp==3.23.0
zstandard==0.25.0
